+++ src/cmd_eval.rs
@@ -9,6 +9,7 @@
 
 use crate::cli::OutputFormat;
 use crate::error::Result;
+use crate::fmt_toml::TomlVersion;
 use crate::pprint::Doc;
 use crate::runtime::Value;
 use crate::source::Span;
@@ -19,7 +20,12 @@ pub fn format_value(format: OutputFormat, value_span: Span, value: &Value) -> Re
         OutputFormat::JsonLines => crate::fmt_json_lines::format_json_lines(value_span, value)?,
         OutputFormat::Raw => crate::fmt_raw::format_raw(value_span, value)?,
         OutputFormat::Rcl => crate::fmt_rcl::format_rcl(value),
-        OutputFormat::Toml => crate::fmt_toml::format_toml(value_span, value)?,
+        OutputFormat::Toml10 => {
+            crate::fmt_toml::format_toml(TomlVersion::Toml10, value_span, value)?
+        }
+        OutputFormat::Toml11 => {
+            crate::fmt_toml::format_toml(TomlVersion::Toml11, value_span, value)?
+        }
         OutputFormat::YamlStream => crate::fmt_yaml_stream::format_yaml_stream(value_span, value)?,
     };
     Ok(result)